+++ printing/lib/src/print_job.dart
@@ -23,7 +23,8 @@ import 'raster.dart';
 /// Represents a print job to communicate with the platform implementation
 class PrintJob {
   /// Create a print job
-  PrintJob({
+  const PrintJob._({
+    required this.index,
     this.onLayout,
     this.onHtmlRendered,
     this.onCompleted,
@@ -43,5 +44,43 @@ class PrintJob {
   final StreamController<PdfRaster>? onPageRasterized;
 
   /// The Job number
-  int? index;
+  final int index;
+}
+
+/// Represents a list of print jobs
+class PrintJobs {
+  /// Create a list print jobs
+  PrintJobs();
+
+  static var _currentIndex = 0;
+
+  final _printJobs = <int, PrintJob>{};
+
+  /// Add a print job to the list
+  PrintJob add({
+    LayoutCallback? onLayout,
+    Completer<Uint8List>? onHtmlRendered,
+    Completer<bool>? onCompleted,
+    StreamController<PdfRaster>? onPageRasterized,
+  }) {
+    final job = PrintJob._(
+      index: _currentIndex++,
+      onLayout: onLayout,
+      onHtmlRendered: onHtmlRendered,
+      onCompleted: onCompleted,
+      onPageRasterized: onPageRasterized,
+    );
+    _printJobs[job.index] = job;
+    return job;
+  }
+
+  /// Retrive an existing job
+  PrintJob? getJob(int index) {
+    return _printJobs[index];
+  }
+
+  /// remove a print job from the list
+  void remove(int index) {
+    _printJobs.remove(index);
+  }
 }